refactor events reducer
