+++ src/EventForm.tsx
@@ -65,7 +65,7 @@ const EventForm: React.FC<{}> = () => {
         <button
           className="btn btn-danger"
           onClick={deleteAllEvents}
-          disabled={state.length === 0}
+          disabled={state.events.length === 0}
         >
           全てのイベントを削除する
         </button>
